refactor(workspace): drop internal.ts re-export barrel
The file existed only to re-export `createKv` / `createReadonlyTable` /
`createTable` from their sibling attach files. It is not in package.json
exports (verified) and has three internal importers (attach-encryption.ts
and two test files). Inline the imports and delete the file.

The barrel claimed to enforce "internal-only" via a separate file, but
nothing actually stopped direct imports from `./attach-table.js` or
`./attach-kv.js`. Internality stays documented per-export in the
`create*` JSDoc.
