+++ packages/workspace/src/document/create-kv.test.ts
@@ -6,8 +6,8 @@ import { expect, test } from 'bun:test';
 import { type } from 'arktype';
 import * as Y from 'yjs';
 import { createEncryptedYkvLww } from '../shared/y-keyvalue/y-keyvalue-lww-encrypted.js';
+import { createKv } from './attach-kv.js';
 import { defineKv } from './define-kv.js';
-import { createKv } from './internal.js';
 import { KV_KEY } from './keys.js';
 
 test('set stores a value that get returns', () => {
